Add comission

+++ src/main/java/main/model/TgUserRepository.java
@@ -3,6 +3,8 @@ package main.model;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface TgUserRepository extends CrudRepository<TgUser, Integer> {
+public interface TgUserRepository extends CrudRepository<TgUser, Long> {
 }
